[WIP] Savestates overhaul
Made the gui look much nicer :).
A few things to note: When savestating the world should be paused aka in tickrate zero. There are a few issues however. The Savestating only starts after closing the savestate menu again. Probably because the server is paused during the LoScreen which would not happen on a dedicated server. It should be   possible to just not pause the server while the LoScreen is opened. Probably by modifying EmptyScreen or smth. Then for Escape Menus, aka it is paused by default, the Savestate menu should not work. Lastly doing this will probably fix the lock screen issue. Another big problem is that the File copy fails because it cannot copy the session.lock. This never unlocks the client and also doesn't add the savestate to the list of states. The Image is also still missing.